// src/interfaces/questionTypes.ts
export type questionTypes =
  | "shorttext"
  | "longtext"
  | "geopoint"
  | "geoshape"
  | "number"
  | "date"
  | "datetime"
  | "time"
  | "audio"
  | "video"
  | "image"
  | "ranking"
  | "rating"
  | "singleselect"
  | "multipleselect"
  | "chooseone"
  | "likert"
  | "group"
  | "cascadingselect"
  | "note"
  | "barcode"
  | "link"
  | "signature"
  | "trigger"
  | "fileupload";